Relocate DB-model and lib documentation to provisional directory
- Move 'DB-model/README.md' to 'zzProvisionalDocs/DB-model/README.md'
- Move 'DB-model/operations-sql.md' to 'zzProvisionalDocs/DB-model/operations-sql.md'
- Move 'DB-model/tables.md' to 'zzProvisionalDocs/DB-model/tables.md'
- Move 'lib/transactions.md' to 'zzProvisionalDocs/lib/transactions.md'

This commit relocates key documentation files from the 'DB-model' and 'lib' directories to the 'zzProvisionalDocs' directory. This move is to clearly designate these files as provisional, ensuring they are not mistakenly used or referenced as finalized or authoritative documents. It is a precautionary step to maintain the integrity and clarity of our documentation, particularly during phases of ongoing review and updates.